Split tokens into a separate module
